Moving 2D unit tests to data_store. About to add testing tolerance to compare to cherab
