+++ raysect/core/math/function/float/function2d/interpolate/tests/test_interpolator_2d.py
@@ -39,561 +39,8 @@ from raysect.core.math.function.float.function2d.interpolate.interpolator2darray
 from raysect.core.math.function.float.function2d.interpolate.tests.scripts.generate_2d_splines import X_LOWER, X_UPPER, \
     NB_XSAMPLES, NB_X, X_EXTRAP_DELTA_MAX, X_EXTRAP_DELTA_MIN, PRECISION, Y_LOWER, Y_UPPER, NB_YSAMPLES, NB_Y, \
     Y_EXTRAP_DELTA_MAX, Y_EXTRAP_DELTA_MIN, EXTRAPOLATION_RANGE, get_extrapolation_input_values
-
-
-class TestInterpolatorLoadValues:
-    def __init__(self):
-        # Define in setup_cubic or setup_linear
-        self.precalc_interpolation = None
-
-        #: array holding precalculated nearest neighbour extrapolation data
-        self.precalc_extrapolation_nearest: np.array = None
-
-        #: array holding precalculated linear extrapolation data
-        self.precalc_extrapolation_linear: np.array = None
-
-        #: array holding precalculated quadratic extrapolation data
-        self.precalc_extrapolation_quadratic: np.array = None
-
-
-class TestInterpolatorLoadNormalValues(TestInterpolatorLoadValues):
-    """
-    Loading normal sized values for a 2D sinc function test.
-
-    These data are saved to 12 significant figures.
-    """
-    def __init__(self):
-        super().__init__()
-        #: data array from a function sampled on self.x. dtype should be np.float64
-        self.data: np.array = np.array(
-            [[7.049456954407E-02, -5.031133752816E-03, -8.474851229653E-02, -7.975908114097E-02,
-              -4.876973734940E-02, -4.876973734940E-02, -7.975908114097E-02, -8.474851229653E-02,
-              -5.031133752816E-03, 7.049456954407E-02],
-             [-5.031133752816E-03, -9.121921863446E-02, -9.251264987298E-04, 1.052139178127E-01,
-              1.283205555674E-01, 1.283205555674E-01, 1.052139178127E-01, -9.251264987298E-04,
-              -9.121921863446E-02, -5.031133752816E-03],
-             [-8.474851229653E-02, -9.251264987298E-04, 1.283205555674E-01, 1.734970013481E-02,
-              -1.140407180451E-01, -1.140407180451E-01, 1.734970013481E-02, 1.283205555674E-01,
-              -9.251264987298E-04, -8.474851229653E-02],
-             [-7.975908114097E-02, 1.052139178127E-01, 1.734970013481E-02, -2.145503300375E-01,
-              -9.241435356589E-02, -9.241435356589E-02, -2.145503300375E-01, 1.734970013480E-02,
-              1.052139178127E-01, -7.975908114097E-02],
-             [-4.876973734940E-02, 1.283205555674E-01, -1.140407180451E-01, -9.241435356589E-02,
-              6.446759109720E-01, 6.446759109720E-01, -9.241435356589E-02, -1.140407180451E-01,
-              1.283205555674E-01, -4.876973734940E-02],
-             [-4.876973734940E-02, 1.283205555674E-01, -1.140407180451E-01, -9.241435356589E-02,
-              6.446759109720E-01, 6.446759109720E-01, -9.241435356589E-02, -1.140407180451E-01,
-              1.283205555674E-01, -4.876973734940E-02],
-             [-7.975908114097E-02, 1.052139178127E-01, 1.734970013481E-02, -2.145503300375E-01,
-              -9.241435356589E-02, -9.241435356589E-02, -2.145503300375E-01, 1.734970013480E-02,
-              1.052139178127E-01, -7.975908114097E-02],
-             [-8.474851229653E-02, -9.251264987298E-04, 1.283205555674E-01, 1.734970013480E-02,
-              -1.140407180451E-01, -1.140407180451E-01, 1.734970013480E-02, 1.283205555674E-01,
-              -9.251264987296E-04, -8.474851229653E-02],
-             [-5.031133752816E-03, -9.121921863446E-02, -9.251264987298E-04, 1.052139178127E-01,
-              1.283205555674E-01, 1.283205555674E-01, 1.052139178127E-01, -9.251264987296E-04,
-              -9.121921863446E-02, -5.031133752816E-03],
-             [7.049456954407E-02, -5.031133752816E-03, -8.474851229653E-02, -7.975908114097E-02,
-              -4.876973734940E-02, -4.876973734940E-02, -7.975908114097E-02, -8.474851229653E-02,
-              -5.031133752816E-03, 7.049456954407E-02]],
-            dtype=np.float64
-        )
-
-    def setup_cubic(self):
-        self.precalc_interpolation = np.array(
-            [[7.049456954407E-02, 1.414501923720E-02, -4.992201892795E-02, -9.006647501703E-02,
-              -7.975908114097E-02, -5.494750801435E-02, -4.489606937545E-02, -5.494750801435E-02,
-              -7.975908114097E-02, -9.006647501703E-02, -4.992201892795E-02, 1.414501923720E-02,
-              7.049456954407E-02],
-             [1.414501923720E-02, -6.879873015242E-02, -6.604387574169E-02, -9.273861048634E-04,
-              7.815453486867E-02, 1.087722976395E-01, 1.179639375774E-01, 1.087722976395E-01,
-              7.815453486867E-02, -9.273861048635E-04, -6.604387574169E-02, -6.879873015242E-02,
-              1.414501923720E-02],
-             [-4.992201892795E-02, -6.604387574169E-02, 7.662704616536E-03, 8.927226959205E-02,
-              8.733637329412E-02, 3.144134389941E-02, 8.046419412818E-03, 3.144134389941E-02,
-              8.733637329412E-02, 8.927226959205E-02, 7.662704616536E-03, -6.604387574169E-02,
-              -4.992201892795E-02],
-             [-9.006647501703E-02, -9.273861048634E-04, 8.927226959205E-02, 9.500724468550E-02,
-              -3.879550824798E-02, -1.262756473836E-01, -1.571105597160E-01, -1.262756473836E-01,
-              -3.879550824798E-02, 9.500724468550E-02, 8.927226959205E-02, -9.273861048634E-04,
-              -9.006647501703E-02],
-             [-7.975908114097E-02, 7.815453486867E-02, 8.733637329412E-02, -3.879550824798E-02,
-              -2.145503300375E-01, -1.226583807439E-01, -7.714735650694E-02, -1.226583807439E-01,
-              -2.145503300375E-01, -3.879550824798E-02, 8.733637329412E-02, 7.815453486867E-02,
-              -7.975908114097E-02],
-             [-5.494750801435E-02, 1.087722976395E-01, 3.144134389941E-02, -1.262756473836E-01,
-              -1.226583807439E-01, 3.691392206491E-01, 5.727262813613E-01, 3.691392206491E-01,
-              -1.226583807439E-01, -1.262756473836E-01, 3.144134389941E-02, 1.087722976395E-01,
-              -5.494750801435E-02],
-             [-4.489606937545E-02, 1.179639375774E-01, 8.046419412818E-03, -1.571105597160E-01,
-              -7.714735650694E-02, 5.727262813613E-01, 8.385571378575E-01, 5.727262813613E-01,
-              -7.714735650694E-02, -1.571105597160E-01, 8.046419412818E-03, 1.179639375774E-01,
-              -4.489606937545E-02],
-             [-5.494750801435E-02, 1.087722976395E-01, 3.144134389941E-02, -1.262756473836E-01,
-              -1.226583807439E-01, 3.691392206491E-01, 5.727262813613E-01, 3.691392206491E-01,
-              -1.226583807439E-01, -1.262756473836E-01, 3.144134389941E-02, 1.087722976395E-01,
-              -5.494750801435E-02],
-             [-7.975908114097E-02, 7.815453486867E-02, 8.733637329412E-02, -3.879550824798E-02,
-              -2.145503300375E-01, -1.226583807439E-01, -7.714735650694E-02, -1.226583807439E-01,
-              -2.145503300375E-01, -3.879550824798E-02, 8.733637329412E-02, 7.815453486867E-02,
-              -7.975908114097E-02],
-             [-9.006647501703E-02, -9.273861048634E-04, 8.927226959205E-02, 9.500724468550E-02,
-              -3.879550824798E-02, -1.262756473836E-01, -1.571105597160E-01, -1.262756473836E-01,
-              -3.879550824798E-02, 9.500724468550E-02, 8.927226959205E-02, -9.273861048634E-04,
-              -9.006647501703E-02],
-             [-4.992201892795E-02, -6.604387574169E-02, 7.662704616536E-03, 8.927226959205E-02,
-              8.733637329412E-02, 3.144134389941E-02, 8.046419412818E-03, 3.144134389941E-02,
-              8.733637329412E-02, 8.927226959205E-02, 7.662704616536E-03, -6.604387574169E-02,
-              -4.992201892795E-02],
-             [1.414501923720E-02, -6.879873015242E-02, -6.604387574169E-02, -9.273861048634E-04,
-              7.815453486867E-02, 1.087722976395E-01, 1.179639375774E-01, 1.087722976395E-01,
-              7.815453486867E-02, -9.273861048634E-04, -6.604387574169E-02, -6.879873015242E-02,
-              1.414501923720E-02],
-             [7.049456954407E-02, 1.414501923720E-02, -4.992201892795E-02, -9.006647501703E-02,
-              -7.975908114097E-02, -5.494750801435E-02, -4.489606937545E-02, -5.494750801435E-02,
-              -7.975908114097E-02, -9.006647501703E-02, -4.992201892795E-02, 1.414501923720E-02,
-              7.049456954407E-02]], dtype=np.float64
-        )
-
-        #: array holding precalculated nearest neighbour extrapolation data
-        self.precalc_extrapolation_nearest: np.array = np.array(
-            [7.049456954407E-02, -4.489606937545E-02, 7.049456954407E-02, -4.489606937545E-02,
-             -4.489606937545E-02, 7.049456954407E-02, -4.489606937545E-02, 7.049456954407E-02], dtype=np.float64
-        )
-
-        #: array holding precalculated linear extrapolation data
-        self.precalc_extrapolation_linear: np.array = np.array(
-            [9.802928389430E-02, -7.659496121465E-02, 9.802928389430E-02, -7.659496121465E-02,
-             -7.659496121465E-02, 9.802928389430E-02, -7.659496121465E-02, 9.802928389430E-02], dtype=np.float64
-        )
-
-    def setup_linear(self):
-        self.precalc_interpolation = np.array(
-            [[7.049456954407E-02, 1.385029207141E-02, -4.488982302467E-02, -8.350115450764E-02,
-              -7.975908114097E-02, -5.651707329729E-02, -4.876973734940E-02, -5.651707329729E-02,
-              -7.975908114097E-02, -8.350115450764E-02, -4.488982302467E-02, 1.385029207141E-02,
-              7.049456954407E-02],
-             [1.385029207141E-02, -4.879157504269E-02, -4.577658518112E-02, -1.668062692568E-03,
-              5.897066807427E-02, 7.777865377223E-02, 8.404798233821E-02, 7.777865377223E-02,
-              5.897066807427E-02, -1.668062692568E-03, -4.577658518112E-02, -4.879157504269E-02,
-              1.385029207141E-02],
-             [-4.488982302467E-02, -4.577658518112E-02, 8.812770983874E-03, 6.309373814419E-02,
-              6.128180897374E-02, 2.067539131431E-02, 7.139918761166E-03, 2.067539131431E-02,
-              6.128180897374E-02, 6.309373814419E-02, 8.812770983874E-03, -4.577658518112E-02,
-              -4.488982302467E-02],
-             [-8.350115450764E-02, -1.668062692568E-03, 6.309373814419E-02, 6.527705442988E-02,
-              -4.062530740828E-02, -9.163192204604E-02, -1.086341269253E-01, -9.163192204604E-02,
-              -4.062530740828E-02, 6.527705442988E-02, 6.309373814419E-02, -1.668062692568E-03,
-              -8.350115450764E-02],
-             [-7.975908114097E-02, 5.897066807427E-02, 6.128180897374E-02, -4.062530740828E-02,
-              -2.145503300375E-01, -1.229483476838E-01, -9.241435356589E-02, -1.229483476838E-01,
-              -2.145503300375E-01, -4.062530740828E-02, 6.128180897374E-02, 5.897066807427E-02,
-              -7.975908114097E-02],
-             [-5.651707329729E-02, 7.777865377223E-02, 2.067539131431E-02, -9.163192204604E-02,
-              -1.229483476838E-01, 3.145654217072E-01, 4.604033448375E-01, 3.145654217072E-01,
-              -1.229483476838E-01, -9.163192204604E-02, 2.067539131431E-02, 7.777865377223E-02,
-              -5.651707329729E-02],
-             [-4.876973734940E-02, 8.404798233821E-02, 7.139918761166E-03, -1.086341269253E-01,
-              -9.241435356589E-02, 4.604033448375E-01, 6.446759109720E-01, 4.604033448375E-01,
-              -9.241435356589E-02, -1.086341269253E-01, 7.139918761166E-03, 8.404798233821E-02,
-              -4.876973734940E-02],
-             [-5.651707329729E-02, 7.777865377223E-02, 2.067539131431E-02, -9.163192204604E-02,
-              -1.229483476838E-01, 3.145654217072E-01, 4.604033448375E-01, 3.145654217072E-01,
-              -1.229483476838E-01, -9.163192204604E-02, 2.067539131431E-02, 7.777865377223E-02,
-              -5.651707329729E-02],
-             [-7.975908114097E-02, 5.897066807427E-02, 6.128180897374E-02, -4.062530740828E-02,
-              -2.145503300375E-01, -1.229483476838E-01, -9.241435356589E-02, -1.229483476838E-01,
-              -2.145503300375E-01, -4.062530740828E-02, 6.128180897374E-02, 5.897066807427E-02,
-              -7.975908114097E-02],
-             [-8.350115450764E-02, -1.668062692568E-03, 6.309373814419E-02, 6.527705442988E-02,
-              -4.062530740828E-02, -9.163192204604E-02, -1.086341269253E-01, -9.163192204604E-02,
-              -4.062530740828E-02, 6.527705442988E-02, 6.309373814419E-02, -1.668062692568E-03,
-              -8.350115450764E-02],
-             [-4.488982302467E-02, -4.577658518112E-02, 8.812770983874E-03, 6.309373814419E-02,
-              6.128180897374E-02, 2.067539131431E-02, 7.139918761166E-03, 2.067539131431E-02,
-              6.128180897374E-02, 6.309373814419E-02, 8.812770983874E-03, -4.577658518112E-02,
-              -4.488982302467E-02],
-             [1.385029207141E-02, -4.879157504269E-02, -4.577658518112E-02, -1.668062692568E-03,
-              5.897066807427E-02, 7.777865377223E-02, 8.404798233821E-02, 7.777865377223E-02,
-              5.897066807427E-02, -1.668062692568E-03, -4.577658518112E-02, -4.879157504269E-02,
-              1.385029207141E-02],
-             [7.049456954407E-02, 1.385029207141E-02, -4.488982302467E-02, -8.350115450764E-02,
-              -7.975908114097E-02, -5.651707329729E-02, -4.876973734940E-02, -5.651707329729E-02,
-              -7.975908114097E-02, -8.350115450764E-02, -4.488982302467E-02, 1.385029207141E-02,
-              7.049456954407E-02]], dtype=np.float64
-        )
-        self.precalc_extrapolation_nearest: np.array = np.array(
-            [7.049456954407E-02, -4.876973734940E-02, 7.049456954407E-02, -4.876973734940E-02,
-             -4.876973734940E-02, 7.049456954407E-02, -4.876973734940E-02, 7.049456954407E-02], dtype=np.float64
-        )
-
-        #: array holding precalculated linear extrapolation data
-        self.precalc_extrapolation_linear: np.array = np.array(
-            [9.802928389430E-02, -8.064599007443E-02, 9.802928389430E-02, -8.064599007443E-02,
-             -8.064599007443E-02, 9.802928389430E-02, -8.064599007443E-02, 9.802928389430E-02], dtype=np.float64
-        )
-
-
-class TestInterpolatorLoadBigValues(TestInterpolatorLoadValues):
-    """
-    Loading big values (10^20 times the original) instead of the original np.sin(x).
-
-    These data are saved to 12 significant figures.
-    """
-    def __init__(self):
-        super().__init__()
-        #: data array from a function sampled on self.x. dtype should be np.float64
-        # self.data: np.array = np.sin(self.x)
-        self.data: np.array = np.array(
-            [[7.049456954407E+18, -5.031133752816E+17, -8.474851229653E+18, -7.975908114097E+18,
-              -4.876973734940E+18, -4.876973734940E+18, -7.975908114097E+18, -8.474851229653E+18,
-              -5.031133752816E+17, 7.049456954407E+18],
-             [-5.031133752816E+17, -9.121921863446E+18, -9.251264987298E+16, 1.052139178127E+19,
-              1.283205555674E+19, 1.283205555674E+19, 1.052139178127E+19, -9.251264987298E+16,
-              -9.121921863446E+18, -5.031133752816E+17],
-             [-8.474851229653E+18, -9.251264987298E+16, 1.283205555674E+19, 1.734970013481E+18,
-              -1.140407180451E+19, -1.140407180451E+19, 1.734970013481E+18, 1.283205555674E+19,
-              -9.251264987298E+16, -8.474851229653E+18],
-             [-7.975908114097E+18, 1.052139178127E+19, 1.734970013481E+18, -2.145503300375E+19,
-              -9.241435356589E+18, -9.241435356589E+18, -2.145503300375E+19, 1.734970013480E+18,
-              1.052139178127E+19, -7.975908114097E+18],
-             [-4.876973734940E+18, 1.283205555674E+19, -1.140407180451E+19, -9.241435356589E+18,
-              6.446759109720E+19, 6.446759109720E+19, -9.241435356589E+18, -1.140407180451E+19,
-              1.283205555674E+19, -4.876973734940E+18],
-             [-4.876973734940E+18, 1.283205555674E+19, -1.140407180451E+19, -9.241435356589E+18,
-              6.446759109720E+19, 6.446759109720E+19, -9.241435356589E+18, -1.140407180451E+19,
-              1.283205555674E+19, -4.876973734940E+18],
-             [-7.975908114097E+18, 1.052139178127E+19, 1.734970013481E+18, -2.145503300375E+19,
-              -9.241435356589E+18, -9.241435356589E+18, -2.145503300375E+19, 1.734970013480E+18,
-              1.052139178127E+19, -7.975908114097E+18],
-             [-8.474851229653E+18, -9.251264987298E+16, 1.283205555674E+19, 1.734970013480E+18,
-              -1.140407180451E+19, -1.140407180451E+19, 1.734970013480E+18, 1.283205555674E+19,
-              -9.251264987296E+16, -8.474851229653E+18],
-             [-5.031133752816E+17, -9.121921863446E+18, -9.251264987298E+16, 1.052139178127E+19,
-              1.283205555674E+19, 1.283205555674E+19, 1.052139178127E+19, -9.251264987296E+16,
-              -9.121921863446E+18, -5.031133752816E+17],
-             [7.049456954407E+18, -5.031133752816E+17, -8.474851229653E+18, -7.975908114097E+18,
-              -4.876973734940E+18, -4.876973734940E+18, -7.975908114097E+18, -8.474851229653E+18,
-              -5.031133752816E+17, 7.049456954407E+18]], dtype=np.float64
-        )
-
-        #: array holding precalculated quadratic extrapolation data
-        self.precalc_extrapolation_quadratic: np.array = np.array(
-            [], dtype=np.float64
-        )
-
-    def setup_cubic(self):
-        self.precalc_interpolation = np.array(
-            [[7.049456954407E+18, 1.414501923720E+18, -4.992201892795E+18, -9.006647501703E+18,
-              -7.975908114097E+18, -5.494750801435E+18, -4.489606937545E+18, -5.494750801435E+18,
-              -7.975908114097E+18, -9.006647501703E+18, -4.992201892795E+18, 1.414501923720E+18,
-              7.049456954407E+18],
-             [1.414501923720E+18, -6.879873015242E+18, -6.604387574169E+18, -9.273861048635E+16,
-              7.815453486867E+18, 1.087722976395E+19, 1.179639375774E+19, 1.087722976395E+19,
-              7.815453486867E+18, -9.273861048635E+16, -6.604387574169E+18, -6.879873015242E+18,
-              1.414501923720E+18],
-             [-4.992201892795E+18, -6.604387574169E+18, 7.662704616536E+17, 8.927226959205E+18,
-              8.733637329412E+18, 3.144134389941E+18, 8.046419412818E+17, 3.144134389941E+18,
-              8.733637329412E+18, 8.927226959205E+18, 7.662704616536E+17, -6.604387574169E+18,
-              -4.992201892795E+18],
-             [-9.006647501703E+18, -9.273861048635E+16, 8.927226959205E+18, 9.500724468550E+18,
-              -3.879550824798E+18, -1.262756473836E+19, -1.571105597160E+19, -1.262756473836E+19,
-              -3.879550824798E+18, 9.500724468550E+18, 8.927226959205E+18, -9.273861048634E+16,
-              -9.006647501703E+18],
-             [-7.975908114097E+18, 7.815453486867E+18, 8.733637329412E+18, -3.879550824798E+18,
-              -2.145503300375E+19, -1.226583807439E+19, -7.714735650694E+18, -1.226583807439E+19,
-              -2.145503300375E+19, -3.879550824798E+18, 8.733637329412E+18, 7.815453486867E+18,
-              -7.975908114097E+18],
-             [-5.494750801435E+18, 1.087722976395E+19, 3.144134389941E+18, -1.262756473836E+19,
-              -1.226583807439E+19, 3.691392206491E+19, 5.727262813613E+19, 3.691392206491E+19,
-              -1.226583807439E+19, -1.262756473836E+19, 3.144134389941E+18, 1.087722976395E+19,
-              -5.494750801435E+18],
-             [-4.489606937545E+18, 1.179639375774E+19, 8.046419412818E+17, -1.571105597160E+19,
-              -7.714735650694E+18, 5.727262813613E+19, 8.385571378575E+19, 5.727262813613E+19,
-              -7.714735650694E+18, -1.571105597160E+19, 8.046419412818E+17, 1.179639375774E+19,
-              -4.489606937545E+18],
-             [-5.494750801435E+18, 1.087722976395E+19, 3.144134389941E+18, -1.262756473836E+19,
-              -1.226583807439E+19, 3.691392206491E+19, 5.727262813613E+19, 3.691392206491E+19,
-              -1.226583807439E+19, -1.262756473836E+19, 3.144134389941E+18, 1.087722976395E+19,
-              -5.494750801435E+18],
-             [-7.975908114097E+18, 7.815453486867E+18, 8.733637329412E+18, -3.879550824798E+18,
-              -2.145503300375E+19, -1.226583807439E+19, -7.714735650694E+18, -1.226583807439E+19,
-              -2.145503300375E+19, -3.879550824798E+18, 8.733637329412E+18, 7.815453486867E+18,
-              -7.975908114097E+18],
-             [-9.006647501703E+18, -9.273861048635E+16, 8.927226959205E+18, 9.500724468550E+18,
-              -3.879550824798E+18, -1.262756473836E+19, -1.571105597160E+19, -1.262756473836E+19,
-              -3.879550824798E+18, 9.500724468550E+18, 8.927226959205E+18, -9.273861048633E+16,
-              -9.006647501703E+18],
-             [-4.992201892795E+18, -6.604387574169E+18, 7.662704616536E+17, 8.927226959205E+18,
-              8.733637329412E+18, 3.144134389941E+18, 8.046419412818E+17, 3.144134389941E+18,
-              8.733637329412E+18, 8.927226959205E+18, 7.662704616536E+17, -6.604387574169E+18,
-              -4.992201892795E+18],
-             [1.414501923720E+18, -6.879873015242E+18, -6.604387574169E+18, -9.273861048634E+16,
-              7.815453486867E+18, 1.087722976395E+19, 1.179639375774E+19, 1.087722976395E+19,
-              7.815453486867E+18, -9.273861048634E+16, -6.604387574169E+18, -6.879873015242E+18,
-              1.414501923720E+18],
-             [7.049456954407E+18, 1.414501923720E+18, -4.992201892795E+18, -9.006647501703E+18,
-              -7.975908114097E+18, -5.494750801435E+18, -4.489606937545E+18, -5.494750801435E+18,
-              -7.975908114097E+18, -9.006647501703E+18, -4.992201892795E+18, 1.414501923720E+18,
-              7.049456954407E+18]], dtype=np.float64
-        )
-        self.precalc_extrapolation_nearest: np.array = np.array(
-            [7.049456954407E+18, -4.489606937545E+18, 7.049456954407E+18, -4.489606937545E+18,
-             -4.489606937545E+18, 7.049456954407E+18, -4.489606937545E+18, 7.049456954407E+18], dtype=np.float64
-        )
-
-        #: array holding precalculated linear extrapolation data
-        self.precalc_extrapolation_linear: np.array = np.array(
-            [9.802928389430E+18, -7.659496121465E+18, 9.802928389430E+18, -7.659496121465E+18,
-             -7.659496121465E+18, 9.802928389430E+18, -7.659496121465E+18, 9.802928389430E+18], dtype=np.float64
-        )
-
-    def setup_linear(self):
-        self.precalc_interpolation = np.array(
-            [[7.049456954407E+18, 1.385029207141E+18, -4.488982302467E+18, -8.350115450764E+18,
-              -7.975908114097E+18, -5.651707329729E+18, -4.876973734940E+18, -5.651707329729E+18,
-              -7.975908114097E+18, -8.350115450764E+18, -4.488982302467E+18, 1.385029207141E+18,
-              7.049456954407E+18],
-             [1.385029207141E+18, -4.879157504269E+18, -4.577658518112E+18, -1.668062692568E+17,
-              5.897066807427E+18, 7.777865377223E+18, 8.404798233821E+18, 7.777865377223E+18,
-              5.897066807427E+18, -1.668062692568E+17, -4.577658518112E+18, -4.879157504269E+18,
-              1.385029207141E+18],
-             [-4.488982302467E+18, -4.577658518112E+18, 8.812770983874E+17, 6.309373814419E+18,
-              6.128180897374E+18, 2.067539131431E+18, 7.139918761166E+17, 2.067539131431E+18,
-              6.128180897374E+18, 6.309373814419E+18, 8.812770983874E+17, -4.577658518112E+18,
-              -4.488982302467E+18],
-             [-8.350115450764E+18, -1.668062692568E+17, 6.309373814419E+18, 6.527705442988E+18,
-              -4.062530740828E+18, -9.163192204604E+18, -1.086341269253E+19, -9.163192204604E+18,
-              -4.062530740828E+18, 6.527705442988E+18, 6.309373814419E+18, -1.668062692568E+17,
-              -8.350115450764E+18],
-             [-7.975908114097E+18, 5.897066807427E+18, 6.128180897374E+18, -4.062530740828E+18,
-              -2.145503300375E+19, -1.229483476838E+19, -9.241435356589E+18, -1.229483476838E+19,
-              -2.145503300375E+19, -4.062530740828E+18, 6.128180897374E+18, 5.897066807427E+18,
-              -7.975908114097E+18],
-             [-5.651707329729E+18, 7.777865377223E+18, 2.067539131431E+18, -9.163192204604E+18,
-              -1.229483476838E+19, 3.145654217072E+19, 4.604033448375E+19, 3.145654217072E+19,
-              -1.229483476838E+19, -9.163192204604E+18, 2.067539131431E+18, 7.777865377223E+18,
-              -5.651707329729E+18],
-             [-4.876973734940E+18, 8.404798233821E+18, 7.139918761166E+17, -1.086341269253E+19,
-              -9.241435356589E+18, 4.604033448375E+19, 6.446759109720E+19, 4.604033448375E+19,
-              -9.241435356589E+18, -1.086341269253E+19, 7.139918761166E+17, 8.404798233821E+18,
-              -4.876973734940E+18],
-             [-5.651707329729E+18, 7.777865377223E+18, 2.067539131431E+18, -9.163192204604E+18,
-              -1.229483476838E+19, 3.145654217072E+19, 4.604033448375E+19, 3.145654217072E+19,
-              -1.229483476838E+19, -9.163192204604E+18, 2.067539131431E+18, 7.777865377223E+18,
-              -5.651707329729E+18],
-             [-7.975908114097E+18, 5.897066807427E+18, 6.128180897374E+18, -4.062530740828E+18,
-              -2.145503300375E+19, -1.229483476838E+19, -9.241435356589E+18, -1.229483476838E+19,
-              -2.145503300375E+19, -4.062530740828E+18, 6.128180897374E+18, 5.897066807427E+18,
-              -7.975908114097E+18],
-             [-8.350115450764E+18, -1.668062692568E+17, 6.309373814419E+18, 6.527705442988E+18,
-              -4.062530740828E+18, -9.163192204604E+18, -1.086341269253E+19, -9.163192204604E+18,
-              -4.062530740828E+18, 6.527705442988E+18, 6.309373814419E+18, -1.668062692568E+17,
-              -8.350115450764E+18],
-             [-4.488982302467E+18, -4.577658518112E+18, 8.812770983874E+17, 6.309373814419E+18,
-              6.128180897374E+18, 2.067539131431E+18, 7.139918761166E+17, 2.067539131431E+18,
-              6.128180897374E+18, 6.309373814419E+18, 8.812770983874E+17, -4.577658518112E+18,
-              -4.488982302467E+18],
-             [1.385029207141E+18, -4.879157504269E+18, -4.577658518112E+18, -1.668062692568E+17,
-              5.897066807427E+18, 7.777865377223E+18, 8.404798233821E+18, 7.777865377223E+18,
-              5.897066807427E+18, -1.668062692568E+17, -4.577658518112E+18, -4.879157504269E+18,
-              1.385029207141E+18],
-             [7.049456954407E+18, 1.385029207141E+18, -4.488982302467E+18, -8.350115450764E+18,
-              -7.975908114097E+18, -5.651707329729E+18, -4.876973734940E+18, -5.651707329729E+18,
-              -7.975908114097E+18, -8.350115450764E+18, -4.488982302467E+18, 1.385029207141E+18,
-              7.049456954407E+18]], dtype=np.float64
-        )
-        self.precalc_extrapolation_nearest: np.array = np.array(
-            [7.049456954407E+18, -4.876973734940E+18, 7.049456954407E+18, -4.876973734940E+18,
-             -4.876973734940E+18, 7.049456954407E+18, -4.876973734940E+18, 7.049456954407E+18], dtype=np.float64
-        )
-
-        #: array holding precalculated linear extrapolation data
-        self.precalc_extrapolation_linear: np.array = np.array(
-            [9.802928389430E+18, -8.064599007443E+18, 9.802928389430E+18, -8.064599007443E+18,
-             -8.064599007443E+18, 9.802928389430E+18, -8.064599007443E+18, 9.802928389430E+18], dtype=np.float64
-        )
-
-
-class TestInterpolatorLoadSmallValues(TestInterpolatorLoadValues):
-    """
-    Loading small values (10^-20 times the original) instead of the original np.sin(x)
-
-    These data are saved to 12 significant figures.
-    """
-    def __init__(self):
-        super().__init__()
-        #: data array from a function sampled on self.x. dtype should be np.float64
-        # self.data: np.array = np.sin(self.x)
-        self.data: np.array = np.array(
-            [[7.049456954407E-22, -5.031133752816E-23, -8.474851229653E-22, -7.975908114097E-22,
-              -4.876973734940E-22, -4.876973734940E-22, -7.975908114097E-22, -8.474851229653E-22,
-              -5.031133752816E-23, 7.049456954407E-22],
-             [-5.031133752816E-23, -9.121921863446E-22, -9.251264987298E-24, 1.052139178127E-21,
-              1.283205555674E-21, 1.283205555674E-21, 1.052139178127E-21, -9.251264987298E-24,
-              -9.121921863446E-22, -5.031133752816E-23],
-             [-8.474851229653E-22, -9.251264987298E-24, 1.283205555674E-21, 1.734970013481E-22,
-              -1.140407180451E-21, -1.140407180451E-21, 1.734970013481E-22, 1.283205555674E-21,
-              -9.251264987298E-24, -8.474851229653E-22],
-             [-7.975908114097E-22, 1.052139178127E-21, 1.734970013481E-22, -2.145503300375E-21,
-              -9.241435356589E-22, -9.241435356589E-22, -2.145503300375E-21, 1.734970013480E-22,
-              1.052139178127E-21, -7.975908114097E-22],
-             [-4.876973734940E-22, 1.283205555674E-21, -1.140407180451E-21, -9.241435356589E-22,
-              6.446759109720E-21, 6.446759109720E-21, -9.241435356589E-22, -1.140407180451E-21,
-              1.283205555674E-21, -4.876973734940E-22],
-             [-4.876973734940E-22, 1.283205555674E-21, -1.140407180451E-21, -9.241435356589E-22,
-              6.446759109720E-21, 6.446759109720E-21, -9.241435356589E-22, -1.140407180451E-21,
-              1.283205555674E-21, -4.876973734940E-22],
-             [-7.975908114097E-22, 1.052139178127E-21, 1.734970013481E-22, -2.145503300375E-21,
-              -9.241435356589E-22, -9.241435356589E-22, -2.145503300375E-21, 1.734970013480E-22,
-              1.052139178127E-21, -7.975908114097E-22],
-             [-8.474851229653E-22, -9.251264987298E-24, 1.283205555674E-21, 1.734970013480E-22,
-              -1.140407180451E-21, -1.140407180451E-21, 1.734970013480E-22, 1.283205555674E-21,
-              -9.251264987296E-24, -8.474851229653E-22],
-             [-5.031133752816E-23, -9.121921863446E-22, -9.251264987298E-24, 1.052139178127E-21,
-              1.283205555674E-21, 1.283205555674E-21, 1.052139178127E-21, -9.251264987296E-24,
-              -9.121921863446E-22, -5.031133752816E-23],
-             [7.049456954407E-22, -5.031133752816E-23, -8.474851229653E-22, -7.975908114097E-22,
-              -4.876973734940E-22, -4.876973734940E-22, -7.975908114097E-22, -8.474851229653E-22,
-              -5.031133752816E-23, 7.049456954407E-22]], dtype=np.float64
-        )
-
-        #: array holding precalculated quadratic extrapolation data
-        self.precalc_extrapolation_quadratic: np.array = np.array(
-            [], dtype=np.float64
-        )
-
-    def setup_cubic(self):
-        self.precalc_interpolation = np.array(
-            [[7.049456954407E-22, 1.414501923720E-22, -4.992201892795E-22, -9.006647501703E-22,
-              -7.975908114097E-22, -5.494750801435E-22, -4.489606937545E-22, -5.494750801435E-22,
-              -7.975908114097E-22, -9.006647501703E-22, -4.992201892795E-22, 1.414501923720E-22,
-              7.049456954407E-22],
-             [1.414501923720E-22, -6.879873015242E-22, -6.604387574169E-22, -9.273861048634E-24,
-              7.815453486867E-22, 1.087722976395E-21, 1.179639375774E-21, 1.087722976395E-21,
-              7.815453486867E-22, -9.273861048636E-24, -6.604387574169E-22, -6.879873015242E-22,
-              1.414501923720E-22],
-             [-4.992201892795E-22, -6.604387574169E-22, 7.662704616536E-23, 8.927226959205E-22,
-              8.733637329412E-22, 3.144134389941E-22, 8.046419412818E-23, 3.144134389941E-22,
-              8.733637329412E-22, 8.927226959205E-22, 7.662704616536E-23, -6.604387574169E-22,
-              -4.992201892795E-22],
-             [-9.006647501703E-22, -9.273861048634E-24, 8.927226959205E-22, 9.500724468550E-22,
-              -3.879550824798E-22, -1.262756473836E-21, -1.571105597160E-21, -1.262756473836E-21,
-              -3.879550824798E-22, 9.500724468550E-22, 8.927226959205E-22, -9.273861048634E-24,
-              -9.006647501703E-22],
-             [-7.975908114097E-22, 7.815453486867E-22, 8.733637329412E-22, -3.879550824798E-22,
-              -2.145503300375E-21, -1.226583807439E-21, -7.714735650694E-22, -1.226583807439E-21,
-              -2.145503300375E-21, -3.879550824798E-22, 8.733637329412E-22, 7.815453486867E-22,
-              -7.975908114097E-22],
-             [-5.494750801435E-22, 1.087722976395E-21, 3.144134389941E-22, -1.262756473836E-21,
-              -1.226583807439E-21, 3.691392206491E-21, 5.727262813613E-21, 3.691392206491E-21,
-              -1.226583807439E-21, -1.262756473836E-21, 3.144134389941E-22, 1.087722976395E-21,
-              -5.494750801435E-22],
-             [-4.489606937545E-22, 1.179639375774E-21, 8.046419412818E-23, -1.571105597160E-21,
-              -7.714735650694E-22, 5.727262813613E-21, 8.385571378575E-21, 5.727262813613E-21,
-              -7.714735650694E-22, -1.571105597160E-21, 8.046419412818E-23, 1.179639375774E-21,
-              -4.489606937545E-22],
-             [-5.494750801435E-22, 1.087722976395E-21, 3.144134389941E-22, -1.262756473836E-21,
-              -1.226583807439E-21, 3.691392206491E-21, 5.727262813613E-21, 3.691392206491E-21,
-              -1.226583807439E-21, -1.262756473836E-21, 3.144134389941E-22, 1.087722976395E-21,
-              -5.494750801435E-22],
-             [-7.975908114097E-22, 7.815453486867E-22, 8.733637329412E-22, -3.879550824798E-22,
-              -2.145503300375E-21, -1.226583807439E-21, -7.714735650694E-22, -1.226583807439E-21,
-              -2.145503300375E-21, -3.879550824798E-22, 8.733637329412E-22, 7.815453486867E-22,
-              -7.975908114097E-22],
-             [-9.006647501703E-22, -9.273861048636E-24, 8.927226959205E-22, 9.500724468550E-22,
-              -3.879550824798E-22, -1.262756473836E-21, -1.571105597160E-21, -1.262756473836E-21,
-              -3.879550824798E-22, 9.500724468550E-22, 8.927226959205E-22, -9.273861048634E-24,
-              -9.006647501703E-22],
-             [-4.992201892795E-22, -6.604387574169E-22, 7.662704616536E-23, 8.927226959205E-22,
-              8.733637329412E-22, 3.144134389941E-22, 8.046419412818E-23, 3.144134389941E-22,
-              8.733637329412E-22, 8.927226959205E-22, 7.662704616536E-23, -6.604387574169E-22,
-              -4.992201892795E-22],
-             [1.414501923720E-22, -6.879873015242E-22, -6.604387574169E-22, -9.273861048634E-24,
-              7.815453486867E-22, 1.087722976395E-21, 1.179639375774E-21, 1.087722976395E-21,
-              7.815453486867E-22, -9.273861048634E-24, -6.604387574169E-22, -6.879873015242E-22,
-              1.414501923720E-22],
-             [7.049456954407E-22, 1.414501923720E-22, -4.992201892795E-22, -9.006647501703E-22,
-              -7.975908114097E-22, -5.494750801435E-22, -4.489606937545E-22, -5.494750801435E-22,
-              -7.975908114097E-22, -9.006647501703E-22, -4.992201892795E-22, 1.414501923720E-22,
-              7.049456954407E-22]], dtype=np.float64
-        )
-        #: array holding precalculated nearest neighbour extrapolation data
-        self.precalc_extrapolation_nearest: np.array = np.array(
-            [7.049456954407E-22, -4.489606937545E-22, 7.049456954407E-22, -4.489606937545E-22,
-             -4.489606937545E-22, 7.049456954407E-22, -4.489606937545E-22, 7.049456954407E-22], dtype=np.float64
-        )
-        #: array holding precalculated linear extrapolation data
-        self.precalc_extrapolation_linear: np.array = np.array(
-            [9.802928389430E-22, -7.659496121465E-22, 9.802928389430E-22, -7.659496121465E-22,
-             -7.659496121465E-22, 9.802928389430E-22, -7.659496121465E-22, 9.802928389430E-22], dtype=np.float64
-        )
-
-    def setup_linear(self):
-        self.precalc_interpolation = np.array(
-            [[7.049456954407E-22, 1.385029207141E-22, -4.488982302467E-22, -8.350115450764E-22,
-              -7.975908114097E-22, -5.651707329729E-22, -4.876973734940E-22, -5.651707329729E-22,
-              -7.975908114097E-22, -8.350115450764E-22, -4.488982302467E-22, 1.385029207141E-22,
-              7.049456954407E-22],
-             [1.385029207141E-22, -4.879157504269E-22, -4.577658518112E-22, -1.668062692568E-23,
-              5.897066807427E-22, 7.777865377223E-22, 8.404798233821E-22, 7.777865377223E-22,
-              5.897066807427E-22, -1.668062692568E-23, -4.577658518112E-22, -4.879157504269E-22,
-              1.385029207141E-22],
-             [-4.488982302467E-22, -4.577658518112E-22, 8.812770983874E-23, 6.309373814419E-22,
-              6.128180897374E-22, 2.067539131431E-22, 7.139918761166E-23, 2.067539131431E-22,
-              6.128180897374E-22, 6.309373814419E-22, 8.812770983874E-23, -4.577658518112E-22,
-              -4.488982302467E-22],
-             [-8.350115450764E-22, -1.668062692568E-23, 6.309373814419E-22, 6.527705442988E-22,
-              -4.062530740828E-22, -9.163192204604E-22, -1.086341269253E-21, -9.163192204604E-22,
-              -4.062530740828E-22, 6.527705442988E-22, 6.309373814419E-22, -1.668062692568E-23,
-              -8.350115450764E-22],
-             [-7.975908114097E-22, 5.897066807427E-22, 6.128180897374E-22, -4.062530740828E-22,
-              -2.145503300375E-21, -1.229483476838E-21, -9.241435356589E-22, -1.229483476838E-21,
-              -2.145503300375E-21, -4.062530740828E-22, 6.128180897374E-22, 5.897066807427E-22,
-              -7.975908114097E-22],
-             [-5.651707329729E-22, 7.777865377223E-22, 2.067539131431E-22, -9.163192204604E-22,
-              -1.229483476838E-21, 3.145654217072E-21, 4.604033448375E-21, 3.145654217072E-21,
-              -1.229483476838E-21, -9.163192204604E-22, 2.067539131431E-22, 7.777865377223E-22,
-              -5.651707329729E-22],
-             [-4.876973734940E-22, 8.404798233821E-22, 7.139918761166E-23, -1.086341269253E-21,
-              -9.241435356589E-22, 4.604033448375E-21, 6.446759109720E-21, 4.604033448375E-21,
-              -9.241435356589E-22, -1.086341269253E-21, 7.139918761166E-23, 8.404798233821E-22,
-              -4.876973734940E-22],
-             [-5.651707329729E-22, 7.777865377223E-22, 2.067539131431E-22, -9.163192204604E-22,
-              -1.229483476838E-21, 3.145654217072E-21, 4.604033448375E-21, 3.145654217072E-21,
-              -1.229483476838E-21, -9.163192204604E-22, 2.067539131431E-22, 7.777865377223E-22,
-              -5.651707329729E-22],
-             [-7.975908114097E-22, 5.897066807427E-22, 6.128180897374E-22, -4.062530740828E-22,
-              -2.145503300375E-21, -1.229483476838E-21, -9.241435356589E-22, -1.229483476838E-21,
-              -2.145503300375E-21, -4.062530740828E-22, 6.128180897374E-22, 5.897066807427E-22,
-              -7.975908114097E-22],
-             [-8.350115450764E-22, -1.668062692568E-23, 6.309373814419E-22, 6.527705442988E-22,
-              -4.062530740828E-22, -9.163192204604E-22, -1.086341269253E-21, -9.163192204604E-22,
-              -4.062530740828E-22, 6.527705442988E-22, 6.309373814419E-22, -1.668062692568E-23,
-              -8.350115450764E-22],
-             [-4.488982302467E-22, -4.577658518112E-22, 8.812770983874E-23, 6.309373814419E-22,
-              6.128180897374E-22, 2.067539131431E-22, 7.139918761166E-23, 2.067539131431E-22,
-              6.128180897374E-22, 6.309373814419E-22, 8.812770983874E-23, -4.577658518112E-22,
-              -4.488982302467E-22],
-             [1.385029207141E-22, -4.879157504269E-22, -4.577658518112E-22, -1.668062692568E-23,
-              5.897066807427E-22, 7.777865377223E-22, 8.404798233821E-22, 7.777865377223E-22,
-              5.897066807427E-22, -1.668062692568E-23, -4.577658518112E-22, -4.879157504269E-22,
-              1.385029207141E-22],
-             [7.049456954407E-22, 1.385029207141E-22, -4.488982302467E-22, -8.350115450764E-22,
-              -7.975908114097E-22, -5.651707329729E-22, -4.876973734940E-22, -5.651707329729E-22,
-              -7.975908114097E-22, -8.350115450764E-22, -4.488982302467E-22, 1.385029207141E-22,
-              7.049456954407E-22]], dtype=np.float64
-        )
-        #: array holding precalculated nearest neighbour extrapolation data
-        self.precalc_extrapolation_nearest: np.array = np.array(
-            [7.049456954407E-22, -4.876973734940E-22, 7.049456954407E-22, -4.876973734940E-22,
-             -4.876973734940E-22, 7.049456954407E-22, -4.876973734940E-22, 7.049456954407E-22], dtype=np.float64
-        )
-        #: array holding precalculated linear extrapolation data
-        self.precalc_extrapolation_linear: np.array = np.array(
-            [9.802928389430E-22, -8.064599007443E-22, 9.802928389430E-22, -8.064599007443E-22,
-             -8.064599007443E-22, 9.802928389430E-22, -8.064599007443E-22, 9.802928389430E-22], dtype=np.float64
-        )
+from raysect.core.math.function.float.function2d.interpolate.tests.data_store.interpololator2d_test_data import \
+    TestInterpolatorLoadBigValues, TestInterpolatorLoadNormalValues, TestInterpolatorLoadSmallValues
 
 
 class TestInterpolators2D(unittest.TestCase):
@@ -820,7 +267,7 @@ class TestInterpolators2D(unittest.TestCase):
 
                 self.run_general_interpolation_tests()
 
-    def run_general_extrapolation_tests(self, gradient_continuity=True, extrapolator_type=''):
+    def run_general_extrapolation_tests(self, gradient_continuity=True, extrapolator_type='', significant_tolerance=None):
         # Test extrapolator out of range, there should be an error raised
         for i in range(len(self.xsamples_out_of_bounds)):
             dict_kwargs_extrapolator_call = {
@@ -851,7 +298,7 @@ class TestInterpolators2D(unittest.TestCase):
                 self.assertAlmostEqual(gradients_lower[0], gradients_lower[1], delta=delta_max_lower)
                 self.assertAlmostEqual(gradients_upper[0], gradients_upper[1], delta=delta_max_upper)
 
-    def run_general_interpolation_tests(self):
+    def run_general_interpolation_tests(self, significant_tolerance=None):
         # Test interpolation against xsample
         for i in range(len(self.xsamples)):
             for j in range(len(self.ysamples)):
